Fix bug and publish new version

+++ dfa_gym/dfa_wrapper.py
@@ -138,7 +138,7 @@ class DFAWrapper(MultiAgentEnv):
             )
 
         if self.dynamic_alphabet:
-            sym2tkn = jax.random.choice(keys[4], self.n_tkns, shape=(self.n_syms,), replace=False, dtype=jnp.int32)
+            sym2tkn = jax.random.choice(keys[4], self.n_tkns, shape=(self.n_syms,), replace=False)
             tkn2sym = (-jnp.ones(self.n_tkns, dtype=jnp.int32)).at[sym2tkn].set(jnp.arange(self.n_syms, dtype=jnp.int32))
         else:
             sym2tkn = jnp.arange(self.n_syms, dtype=jnp.int32)
